modified:   11.cpp 	modified:   278.cpp 	modified:   283.cpp 	modified:   290.cpp 	modified:   292.cpp 	modified:   3.cpp 	modified:   342.cpp 	modified:   345.cpp 	modified:   349.cpp 	modified:   367.cpp 	modified:   371.cpp 	modified:   374.cpp 	modified:   400.cpp 	modified:   455.cpp 	modified:   475.cpp

+++ 400.cpp
@@ -3,7 +3,7 @@ class Solution
 public:
     long getIndex(long n, long digits)
     {
-        int number = n/digits + (digits == 1 ? 0 : static_cast<int>(pow(10, digits-1)));
+        int number = n/digits + (digits == 1 ? 0 : static_cast<long>(pow(10, digits-1)));
         for (int i = 1; i < digits - n%digits; ++i)
         {
             number /= 10;
